backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.30,<1
psycopg[binary]>=3.2,<4
python-dotenv>=1,<2
